snabb: Simplify "max packets per batch" mechanism
Remove the extra test for how many packets are available. This seemed
to help performance in earlier tests but the results are not actually
reproducible enough to be confident that this is faster vs slower.

Replace with a static limit on how many packets can be processed at
one time.