snabb: Use simple accumulator variable in test code
The return values of fastnetmon_parse_pkt() are now added up in a
variable called 'accumulator'. This is done so that the return value
is really used and GCC cannot treat the parsing as dead code when
inlining and optimizing at -O3.

(Probably the previous version of this code was meaningless to
benchmark due to excessive optimization by GCC. Hopefully it is better
now.)